+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Better CICD
+- `setSlowThreshold` to mutate slow-classification cutoff at runtime
 
 ## [0.1.0] - 2026-05-18
 ### Added